Fix implementation of MongoDbExtension#isUnitTestDatabaseForThisService
The original implementation (which did NOT have actual unit tests) does
not work if the service name is truncated. Changed the behavior to check
if the given databaseName argument (after stripping timestamp) equals
the database name (again after stripping the timestamp) of the current
test database this extension is using. This is a more strict check and
is also more correct, since it compares the exact database names
minus the timestamp (which we clearly cannot include in the comparison
since they will always be different)