Switch to build on Java 17 and Temurin (#303)
* Switch to build on Java 17 and Temurin

* Add lovely hack to allow this to work the same on JDK 11-16 and 17+

The hack in Fixtures#fixture(String, Charset) adds a new catch block
that catches Error. When the Error's cause is a CharacterCodingException
we throw an UncheckedIOException with the CharacterCodingException as
its cause. This replicates the behavior from JDK 11-16 when running on a
17 JDK and presumably 18, 19, etc. I am pretty sure this is a regression
in the JDK's Files#readString method because its Javadoc says that an
IOException is thrown for malformed input or unmappable byte sequence.

But, since I certainly can't control the JDK and don't even know how to
go about reporting or asking whether this is a real problem, for now
I've just hacked around it.

Since our code coverage is only run on JDK 11 at present, I also
extracted a separate method that encapsulates this special handling code
in order to fully test it on JDKs before 17.

* Just call getCause() on the Error object

This avoids generating the stack trace as a String, which is what
ThrowableInfo#of will do. Since we only need the cause of the Throwable,
we can just get the cause directly from the Error object. Not sure why
I initially used ThrowableInfo.