Handle nested test classes in H2FileBasedDatabaseExtension (#334)
* Handle nested test classes in H2FileBasedDatabaseExtension

* Change H2FileBasedDatabaseExtension#beforeAll so that it only creates
   the test database at the top-level, i.e. not when it encounters
   a nested test class.
* Similarly, update  H2FileBasedDatabaseExtension#beforeAll to only
  delete the test database at the top-level

Fixes #333

* Extract JupiterHelpers utility class

Allows testing of the fallback mechanism used in the code
to detect whether a test class is annotated with Nested. Also
allows sharing this code between our extensions (and we have
others that could use the Nested class check, e.g. possibily the
CuratorTestingServerExtension).

* Simplify...if there is a database, just return

If a database already exists, just return since almost
certainly in a nested test class. Warning about a
condition that we don't expect to ever happen does
not add value. If the database does not exist, then
create it and store it.

* Fix the javadoc that is now incorrect.

* Make the afterAll javadoc more clear (I think)

* Remove unused import

* Minor cleanup

* Fix imports in JupiterHelpers (using IntelliJ)
* Fix grammar errors in JupiterHelpers
* Suppress some SQL warnings in the H2FileBasedDatabaseExtension tests
  (these are IntelliJ-specific suppression values)