Fix "potential databaser esource leak" in tests
I'm pretty sure this would not be an issue even in production code, but
move the Statement creation into the try-with-resources declaration
to ensure closure. I think it would not really be an issue because the
Connection is closed automatically, which I belive also closes any
associated Statememts. But, since I am not 100% sure, just make
sure.