Fix CodeQL workflow to work with Lombok (#365)
Updates to the CodeQL workflow (codeql.yml):

* Fix CodeQL + Lokbok by adding delombok
* Remove extraneous comments
* Cache maven artifacts
* Allow manual triggering (workflow_dispatch)

Additional changes were needed to make CodeQL analysis run without errors.
Specifically, replaced Lombok Delegate annotations in H2FileBasedDatabaseExtension
and H2LiquibaseExtension with "real" methods that perform the delegate the
"old fashioned" way. Otherwise the de-lomboked code did not compile because
the methods that should have been created by the Delegate annotation were not
in the (de-lomboked) source files. I am not sure whether this is something to
do with the way the advanced-security/delombok action works or some other reason.
In any case, since the Delegate annotation is experimental in Lombok, and its description=
web page says it has a "negative" status and it will not move out of experimental status
"anytime soon" and also that "support for this feature may be dropped if future versions
of javac or ecj make it difficult to continue to maintain the feature". So, the safe thing to
do is just write the (simple) code that it produced, so that the CodeQL workflow can
run without failing.

Additional fixes caught by CodeQL:

* Fix erroneous Javadoc in ParameterizedTestHelper
* Fix "potential database resource leak" in H2FileBasedDatabaseExtensionRegisterTest
  and H2FileBasedDatabaseExtensionExtendWithTest
* Catch NumberFormatException in MongoTestProperties#extractDatabaseTimestamp
  and re-throw as an IllegalArgumentException.

The above "additional fixes" were captured as issue #366.

Last, dismissed a whole bunch of false positive CodeQL warnings and notes. These were
all caused (I am about 99.99999% sure anyway) due to Lombok-generated code. Not that
the Lombok code is wrong, rather I think in some cases CodeQL doesn't understand Lombok
while in other cases CodeQL was just wrong. For example, CodeQL does not understand that
a Lombok Singular annotation on a List field ensures it won't be null. It also said there was
a boxed primitive comparison on an int field, which was inside a class annotated with Lombok's
Value annotation.

Closes #366