Temporary: re-enable H2LiquibaseExtensionWithJdbiExtensionTest...
...the previous run passed with only H2LiquibaseExtensionTest. I expect
this one to FAIL due to something being cached somewhere and maybe the
way Maven is being executed, i.e. in the same JVM such that static
vars are being retained. AbstractUpdateCommandStep#isUpToDateFastCheck
does a check if the DB is up-to-date, and checks a cache backed by a
static Map. So, if it is the same JVM and the cache key, then maybe
that's why it "thinks" things are uo-to-date!