Return ConstraintViolation objects from validation test helpers
Change void methods in SoftValidationTestAssertions and
ValidationTestHelper to return a Set of ConstraintViolation objects
or a single ConstraintViolation object, as appropriate.

The void methods in ValidationTestHelper that validate there are no
violations, and which do not use SoftAssertions, remain as void. This
is simply because when the assertion passes, the set of constraint
violations will always be empty, so returning it provides no value.

Perhaps surprisingly at first, the validation methods using
SoftAssertions always have a return value. Because of how soft
assertions work (i.e. they collect errors and do not fail
immediately), you cannot know exactly what the validation results
are, so we therefore always have to return the violations for
inspection. For example, validating that there are two errors
on an object using soft assertions when there are actually none,
will result in a Set of ConstraintViolation objects being returned
that is empty. So, any further (soft or not) assertions performed
on the returned Set will also result in errors. Or, validating that
there are no errors on an object, again using soft assertions, when
there are actually three errors, will return a Set of
ConstraintViolation that contains those three errors. When using
soft assertions, the return values will be most useful to perform
additional (soft) assertions on the ConstraintViolation objects, but
they may also be useful during debugging such as when you aren't
expecting any validation errors but there are some.

Closes #403